safely remove Toolbar
