+++ packages/@tinacms/toolkit/src/packages/react-forms/index.ts
@@ -16,4 +16,4 @@ limitations under the License.
 
 */
 
-export * from './CreateContentMenu'
+export * from './FormModal'
