--- packages/@tinacms/toolkit/src/packages/react-toolbar/toolbar.ts
@@ -1,47 +1,0 @@
-/**
-
-Copyright 2021 Forestry.io Holdings, Inc.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-
-*/
-
-export interface ToolbarStateOptions {
-  buttons?: ToolbarButtons
-}
-
-export interface ToolbarButtons {
-  save: string
-  reset: string
-}
-
-export class ToolbarState {
-  buttons: ToolbarButtons = {
-    save: 'Save',
-    reset: 'Reset',
-  }
-
-  constructor(options: ToolbarStateOptions = {}) {
-    /*
-     ** TODO: Do we want to handle buttons
-     ** the same as sidebar? Or make them plugins?
-     ** maybe remove these
-     */
-    if (options.buttons?.save) {
-      this.buttons.save = options.buttons.save
-    }
-    if (options.buttons?.reset) {
-      this.buttons.reset = options.buttons.reset
-    }
-  }
-}
